requirements.txt
torch>=0.4.0
torchvision
matplotlib
numpy
scipy
pillow
urllib3
scikit-image

Only for training, need to install pytorch-msssim
pip install pytorch-msssim

